+++ configs/typescript.js
@@ -1,3 +1,5 @@
+const coreRules = require('./core').rules;
+
 module.exports = {
   overrides: [
     {
@@ -48,6 +50,10 @@ module.exports = {
             },
           },
         ],
+
+        'no-unused-vars': 'off', // note you must disable the base rule as it can report incorrect errors
+        '@typescript-eslint/no-unused-vars': coreRules['no-unused-vars'],
+
         '@typescript-eslint/no-unnecessary-type-arguments': 'warn',
         '@typescript-eslint/consistent-type-definitions': ['warn', 'type'],